since passing gateway checks takes some time, I moved the laoding animation start to before any checks. Separated all the checks to a CheckManager class
