+++ YoCode/Program.cs
@@ -1,7 +1,6 @@
 ﻿using System.Linq;
 using System.Collections.Generic;
 using System.Threading;
-using System;
 
 namespace YoCode
 {
@@ -49,47 +48,6 @@ namespace YoCode
 
             showLoadingAnim = !result.NoLoadingScreen;
 
-            var evidenceList = new List<FeatureEvidence>();
-
-            var projectRunner = PassGatewayChecks(dir, evidenceList);
-            if (projectRunner == null)
-            {
-                compositeOutput.PrintFinalResults(evidenceList, 0);
-                return;
-            }
-
-            evidenceList = PerformChecks(dir, parameters, projectRunner);
-            compositeOutput.PrintFinalResults(evidenceList.OrderBy(a => a.FeatureTitle),
-                new Results(evidenceList, isJunior ? TestType.Junior : TestType.Original).FinalScore);
-        }
-
-        private static ProjectRunner PassGatewayChecks(IPathManager dir, ICollection<FeatureEvidence> evidenceList)
-        {
-            var fileCheck = new FileChangeFinder(dir.ModifiedTestDirPath);
-            evidenceList.Add(fileCheck.FileChangeEvidence);
-            if (fileCheck.FileChangeEvidence.FeatureFailed)
-            {
-                return null;
-            }
-
-            var projectBuilder = new ProjectBuilder(dir.ModifiedTestDirPath, new FeatureRunner());
-            evidenceList.Add(projectBuilder.ProjectBuilderEvidence);
-            if (projectBuilder.ProjectBuilderEvidence.FeatureFailed)
-            {
-                return null;
-            }
-
-            var projectRunner = new ProjectRunner(dir.ModifiedTestDirPath, new FeatureRunner());
-            ConsoleCloseHandler.StartHandler(projectRunner);
-            projectRunner.Execute();
-            evidenceList.Add(projectRunner.ProjectRunEvidence);
-            return projectRunner.ProjectRunEvidence.FeatureFailed ? null : projectRunner;
-        }
-
-        private static List<FeatureEvidence> PerformChecks(IPathManager dir, RunParameterChecker p, ProjectRunner projectRunner)
-        {
-            var checkList = new List<FeatureEvidence>();
-
             var workThreads = new List<Thread>();
 
             if (showLoadingAnim)
@@ -102,49 +60,22 @@ namespace YoCode
                 loadingThread.Start();
             }
 
-            // CodeCoverage check
-            var codeCoverage = new Thread(() =>
+            var evidenceList = new List<FeatureEvidence>();
+
+            var checkManager = new CheckManager(dir, workThreads, isJunior);
+
+            var projectRunner = checkManager.PassGatewayChecks(evidenceList);
+
+            if (projectRunner == null)
             {
-                checkList.Add(new CodeCoverageCheck(p.DotCoverDir, dir.ModifiedTestDirPath, new FeatureRunner()).CodeCoverageEvidence);
-            });
-            workThreads.Add(codeCoverage);
-            codeCoverage.Start();
+                compositeOutput.PrintFinalResults(evidenceList, 0);
+                return;
+            }
 
-            // Duplication check
-            var dupFinderThread = new Thread(() =>
-            {
-                checkList.Add(new DuplicationCheck(dir, new DupFinder(p.CMDToolsPath), isJunior).DuplicationEvidence);
-            });
-            workThreads.Add(dupFinderThread);
-            dupFinderThread.Start();
+            evidenceList = checkManager.PerformChecks(parameters, projectRunner);
 
-            // UI test
-            var modifiedHtmlFiles = dir.GetFilesInDirectory(dir.ModifiedTestDirPath, FileTypes.html).ToList();
-
-            checkList.Add(new UICheck(modifiedHtmlFiles, UIKeywords.UNIT_KEYWORDS).UIEvidence);
-
-            // Git repo used
-            checkList.Add(new GitCheck(dir.ModifiedTestDirPath).GitEvidence);
-
-            // Unit test test
-            checkList.Add(new TestCountCheck(dir.ModifiedTestDirPath, new FeatureRunner()).UnitTestEvidence);
-
-            //Front End Check
-            checkList.Add(new FrontEndCheck(projectRunner.GetPort(), UIKeywords.UNIT_KEYWORDS).FrontEndEvidence);
-
-            var ucc = new UnitConverterCheck(projectRunner.GetPort());
-
-            // Unit converter test
-            checkList.Add(ucc.UnitConverterCheckEvidence);
-
-            checkList.Add(ucc.BadInputCheckEvidence);
-
-            LoadingAnimation.LoadingFinished = true;
-            workThreads.ForEach(a => a.Join());
-            projectRunner.KillProject();
-
-            projectRunner.ReportLefOverProcess();
-            return checkList;
+            compositeOutput.PrintFinalResults(evidenceList.OrderBy(a => a.FeatureTitle),
+                new Results(evidenceList, isJunior ? TestType.Junior : TestType.Original).FinalScore);
         }
     }
 }
